Move transfromProperties and getPropertyValue into model/utils.
