+++ src/model/utils/utils.ts
@@ -2,6 +2,7 @@ import UniProtKBEntryConfig from '../../view/uniprotkb/UniProtEntryConfig';
 import { UniProtkbUIModel } from '../uniprotkb/UniProtkbConverter';
 import { GeneNamesData } from '../uniprotkb/sections/NamesAndTaxonomyConverter';
 import { uniq } from '../../utils/utils';
+import { Property, PropertyKey } from '../types/modelTypes';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export const hasContent = (obj: any) => {
@@ -43,4 +44,23 @@ export const flattenGeneNameData = (geneNamesData: GeneNamesData) => {
     }
   );
   return uniq(geneNames);
+};
+
+export const transfromProperties = (properties: Property[]) => {
+  const o: { [key: string]: string } = {};
+  properties.forEach(({ key, value }) => {
+    if (key && value) {
+      o[key] = value;
+    }
+  });
+  return o;
+};
+
+// This function is useful because our API returns arrays of objects of shape: { key: x, value: y}
+export const getPropertyValue = (
+  properties: Property[],
+  propertyKey: PropertyKey
+) => {
+  const found = properties.find(({ key }) => key === propertyKey);
+  return found ? found.value : null;
 };
